# dist/public/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "mongoose.Types.ObjectId": {
                "type": "string"
            },
...
            "mongoose.FlattenMaps__name-string--symbol-string--icon-string--rate-number--vipRate-number--networks_58__name-string--address-string--icon_63_-string_-Array__": {
                "properties": {},
                "type": "object"
            },
            "mongoose.FlattenMaps_Asset-and-__id-mongoose.Types.ObjectId_-and-___v-number__": {
                "properties": {},
                "type": "object"
            },
            "IAddAssetInput": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "symbol": {
                        "type": "string"
                    },
                    "rate": {
                        "type": "number",
                        "format": "double"
                    },
                    "vipRate": {
                        "type": "number",
                        "format": "double"
                    },
                    "networks": {
                        "items": {
                            "properties": {
                                "icon": {
                                    "type": "string"
                                },
                                "address": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "address",
                                "name"
                            ],
                            "type": "object"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "name",
                    "icon",
                    "symbol",
                    "rate",
                    "networks"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IUpdateAssetInput": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "symbol": {
                        "type": "string"
                    },
                    "rate": {
                        "type": "number",
                        "format": "double"
                    },
                    "vipRate": {
                        "type": "number",
                        "format": "double"
                    },
                    "networks": {
                        "items": {
                            "properties": {
                                "icon": {
                                    "type": "string"
                                },
                                "address": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "address",
                                "name"
                            ],
                            "type": "object"
                        },
                        "type": "array"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/assets": {
            "get": {
                "operationId": "GetAllAssets",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/mongoose.FlattenMaps__name-string--symbol-string--icon-string--rate-number--vipRate-number--networks_58__name-string--address-string--icon_63_-string_-Array__"
                                                    },
                                                    {
                                                        "properties": {
                                                            "_id": {
                                                                "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                                                            }
                                                        },
                                                        "required": [
                                                            "_id"
                                                        ],
                                                        "type": "object"
                                                    },
                                                    {
                                                        "properties": {
                                                            "__v": {
                                                                "type": "number",
                                                                "format": "double"
                                                            }
                                                        },
                                                        "required": [
                                                            "__v"
                                                        ],
                                                        "type": "object"
                                                    }
                                                ]
                                            },
                                            "type": "array"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Assets"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "CreateAsset",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/mongoose.FlattenMaps_Asset-and-__id-mongoose.Types.ObjectId_-and-___v-number__"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Assets"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IAddAssetInput"
                            }
                        }
                    }
                }
            }
        },
        "/assets/{id}": {
            "get": {
                "operationId": "GetAsset",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/mongoose.FlattenMaps__name-string--symbol-string--icon-string--rate-number--vipRate-number--networks_58__name-string--address-string--icon_63_-string_-Array__"
                                                },
                                                {
                                                    "properties": {
                                                        "_id": {
                                                            "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                                                        }
                                                    },
                                                    "required": [
                                                        "_id"
                                                    ],
                                                    "type": "object"
                                                },
                                                {
                                                    "properties": {
                                                        "__v": {
                                                            "type": "number",
                                                            "format": "double"
                                                        }
                                                    },
                                                    "required": [
                                                        "__v"
                                                    ],
                                                    "type": "object"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Assets"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "patch": {
                "operationId": "UpdateAsset",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/mongoose.FlattenMaps_Asset-and-__id-mongoose.Types.ObjectId_-and-___v-number__"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Assets"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IUpdateAssetInput"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteAsset",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Assets"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "info": {
        "license": {
            "name": "ISC"
        },
        "contact": {},
        "title": "MD Exchange API",
        "description": "API for managing MD Exchange operations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "localhost dev"
        }
    ]
}